requirements.txt
Flask==3.1.0
Flask-Cors==5.0.0
Werkzeug==3.1.3
newspaper3k==0.2.8
groq==0.18.0
lxml==5.3.1
lxml_html_clean==0.4.1